docs/azure/requirements.txt
azure-ai-textanalytics==5.3.0
azure-data-tables==12.4.4
python-dotenv[cli]
